// tools/llm-sequential-upgrade/sequential-upgrade/sequential-upgrade-20260403/spacetime/results/chat-app-20260403-131930/level-3/client/src/module_bindings/types.ts
// THIS FILE IS AUTOMATICALLY GENERATED BY SPACETIMEDB. EDITS TO THIS FILE
// WILL NOT BE SAVED. MODIFY TABLES IN YOUR MODULE SOURCE CODE INSTEAD.

/* eslint-disable */
/* tslint:disable */
import {
  TypeBuilder as __TypeBuilder,
  t as __t,
  type AlgebraicTypeType as __AlgebraicTypeType,
  type Infer as __Infer,
} from "spacetimedb";

export const Message = __t.object("Message", {
  id: __t.u64(),
  roomId: __t.u64(),
  sender: __t.identity(),
  text: __t.string(),
  sentAt: __t.timestamp(),
  expiresAtMicros: __t.u64(),
});
export type Message = __Infer<typeof Message>;

export const ReadReceipt = __t.object("ReadReceipt", {
  id: __t.u64(),
  roomId: __t.u64(),
  userIdentity: __t.identity(),
  lastReadMessageId: __t.u64(),
});
export type ReadReceipt = __Infer<typeof ReadReceipt>;

export const Room = __t.object("Room", {
  id: __t.u64(),
  name: __t.string(),
  createdBy: __t.identity(),
  createdAt: __t.timestamp(),
});
export type Room = __Infer<typeof Room>;

export const RoomMember = __t.object("RoomMember", {
  id: __t.u64(),
  roomId: __t.u64(),
  userIdentity: __t.identity(),
});
export type RoomMember = __Infer<typeof RoomMember>;

export const ScheduledMessage = __t.object("ScheduledMessage", {
  scheduledId: __t.u64(),
  scheduledAt: __t.scheduleAt(),
  roomId: __t.u64(),
  sender: __t.identity(),
  text: __t.string(),
});
export type ScheduledMessage = __Infer<typeof ScheduledMessage>;

export const TypingIndicator = __t.object("TypingIndicator", {
  id: __t.u64(),
  roomId: __t.u64(),
  userIdentity: __t.identity(),
  updatedAt: __t.timestamp(),
});
export type TypingIndicator = __Infer<typeof TypingIndicator>;

export const User = __t.object("User", {
  identity: __t.identity(),
  name: __t.string(),
  online: __t.bool(),
});
export type User = __Infer<typeof User>;